pretty functions relocated to parson.c
